Added list handling and rd_string_printf().
This functionality adds tests too.

+++ test/run-all.sh
@@ -4,14 +4,17 @@ run_test() {
 	echo "# Running $1..."
 	./$1 > $1.got 2>&1
 
-	diff -up $1.got $1.expected
-
-	# cleanup if good
-	if [ $? -eq 0 ]; then
-		cat $1.got
-		rm $1.got
-	else 
-		export RSSGET_TESTS_STATUS=1
+	if [ -f $1.expected ]; then
+		diff -up $1.got $1.expected
+		# cleanup if good
+		if [ $? -eq 0 ]; then
+			cat $1.got
+			rm $1.got
+		else
+			export RSSGET_TESTS_STATUS=1
+		fi
+	else
+		echo -e "\e[1;31m*** Missing ***\e[0m $1.expected - skipping '$1' test..."
 	fi
 }
